feat(api): return empty ccd for cw
refs. metriport/metriport-internal#2240
