+++ packages/core/src/external/carequality/dq/process-inbound-dq.ts
@@ -1,11 +1,7 @@
 import { InboundDocumentQueryReq, InboundDocumentQueryResp } from "@metriport/ihe-gateway-sdk";
-import { executeWithNetworkRetries } from "@metriport/shared";
-import axios from "axios";
-import { CCD_SUFFIX, createUploadFilePath } from "../../../domain/document/upload";
+import { ensureCcdExists } from "../../../shareback/ensure-ccd-exists";
 import { getMetadataDocumentContents } from "../../../shareback/metadata/get-metadata-xml";
-import { Config } from "../../../util/config";
 import { out } from "../../../util/log";
-import { S3Utils } from "../../aws/s3";
 import {
   IHEGatewayError,
   XDSRegistryError,
@@ -15,14 +11,8 @@ import {
 import { validateBasePayload } from "../shared";
 import { decodePatientId } from "./utils";
 
-const region = Config.getAWSRegion();
-const s3Utils = new S3Utils(region);
-const api = axios.create();
-const bucket = Config.getMedicalDocumentsBucketName();
-
 export async function processInboundDq(
-  payload: InboundDocumentQueryReq,
-  apiUrl: string
+  payload: InboundDocumentQueryReq
 ): Promise<InboundDocumentQueryResp> {
   try {
     validateBasePayload(payload);
@@ -34,28 +24,9 @@ export async function processInboundDq(
     const { cxId, id: patientId } = id_pair;
     const { log } = out(`Inbound DQ: ${cxId}, patientId: ${patientId}`);
 
-    const destinationKey = createUploadFilePath(cxId, patientId, `${CCD_SUFFIX}.xml`);
-    const ccdExists = await s3Utils.fileExists(bucket, destinationKey);
-    if (!ccdExists) {
-      log("No CCD found. Let's trigger generating one.");
-      const queryParams = {
-        cxId,
-        patientId,
-      };
-      const params = new URLSearchParams(queryParams).toString();
+    await ensureCcdExists({ cxId, patientId, log });
 
-      executeWithNetworkRetries(async () => api.post(`${apiUrl}/internal/docs/ccd?${params}`), {
-        log,
-      });
-      await executeWithNetworkRetries(
-        async () => await api.post(`${apiUrl}/internal/docs/empty-ccd?${params}`),
-        { log }
-      );
-
-      log("CCD generated. Fetching the document contents");
-    }
-
-    const metadataDocumentContents = await getMetadataDocumentContents(cxId, patientId, false);
+    const metadataDocumentContents = await getMetadataDocumentContents(cxId, patientId);
     const response: InboundDocumentQueryResp = {
       id: payload.id,
       patientId: payload.patientId,
